Fix IntegrationFileUploadField to be usable for csv & other filetypes
