+++ src/ui/common/src/components/integrations/dialogs/csvDialog.tsx
@@ -1,8 +1,9 @@
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
+import { DataGrid } from '@mui/x-data-grid';
 import React, { useEffect, useState } from 'react';
 
-import { CSVConfig } from '../../../utils/integrations';
+import { CSVConfig, FileData } from '../../../utils/integrations';
 import { IntegrationFileUploadField } from './IntegrationFileUploadField';
 import { IntegrationTextInputField } from './IntegrationTextInputField';
 
@@ -58,6 +59,37 @@ export const CSVDialog: React.FC<Props> = ({ setDialogConfig, setErrMsg }) => {
               reader.readAsText(file);
             }
           }
+        }}
+        displayFile={(file: FileData) => {
+          const allRows = file.data.split(/\r?\n/);
+          const parsedHeader = ['id'];
+          parsedHeader.push(...allRows[0].split(/,/));
+          const width = 25;
+          const parsedColumns = parsedHeader.map((headerName) => {
+            return {
+              field: headerName,
+              headerName: headerName,
+              width: width * headerName.length,
+            };
+          });
+          const parsedRows = allRows.slice(1).map((line, id) => {
+            const row = line.split(/,/);
+            const parsedRow = { id: id };
+            parsedHeader.forEach(
+              (headerName, i) => (parsedRow[headerName] = row[i])
+            );
+            return parsedRow;
+          });
+
+          return (
+            <DataGrid
+              rows={parsedRows}
+              columns={parsedColumns}
+              pageSize={5}
+              rowsPerPageOptions={[5]}
+              disableSelectionOnClick
+            />
+          );
         }}
         onReset={(_) => {
           setName('');
